requirements.txt
transformers>=4.36
torch>=2.1
datasets
scikit-learn
pytesseract
Pillow
beautifulsoup4
openai
langchain
tqdm
fpdf
python-dotenv
streamlit
altair
schedule
transformers>=4.41.1
accelerate>=0.26.0
datasets
scikit-learn
torch>=2.0.0
tqdm
openai
pydantic
python-dotenv
email-validator
matplotlib
sentencepiece
tiktoken
datasets
python-dotenv
pytest>=7.0.0
fpdf2
openai>=1.12.0
tabulate>=0.9.0
pytest-html
rich>=13.0.0
tenacity>=8.0.1
